Wait for Cassandra server to start before running the tests
Latest version fo Cassandra docker image uses Cassandra 4. This version takes longer
to start the service, possibly due to this issue:

https://issues.apache.org/jira/browse/CASSANDRA-15850

That is causing some builds to fail since the Cassandra server may no be up when the
tests starts. The fix then is to have a healthcheck on docker compose and only run
the tests after the Cassadra server is healthy.