244: Setup for pekko: Remove the deploy-wip workflow and modify the deploy-site script
The gustav VM was used for test purposes. In akka there was  long-lifetime akka.io website that main pipeline deploy-wip workflows would continuously deploy to. This will not be the case for pekko-platform guides so this is removed here.

I have kept the deploy-site script in case anyone wants to host their own local VM TEST_DEPLOYMENT_VM and deploy the site onto that.